(feat): finish handling of `nullable-string-array`
