+++ src/anndata/experimental/backed/_lazy_arrays.py
@@ -144,6 +144,9 @@ class MaskedArray(BackendArray, Generic[K]):
             extension_array = pd.arrays.IntegerArray(values, mask=mask)
         elif self._dtype_str == "nullable-boolean":
             extension_array = pd.arrays.BooleanArray(values, mask=mask)
+        elif self._dtype_str == "nullable-string-array":
+            values[mask] = None
+            extension_array = pd.array(values, dtype="string")
         else:
             msg = f"Invalid dtype_str {self._dtype_str}"
             raise RuntimeError(msg)
@@ -158,6 +161,8 @@ class MaskedArray(BackendArray, Generic[K]):
             ).dtype
         elif self._dtype_str == "nullable-boolean":
             return pd.BooleanDtype()
+        elif self._dtype_str == "nullable-string-array":
+            return pd.StringDtype()
         msg = f"Invalid dtype_str {self._dtype_str}"
         raise RuntimeError(msg)
